requirements.txt
boto3
python-dotenv
schedule
email-validator
python-magic
